Changed optimizer class from being a factory for the algorithm class to being the vessel for actually performing optimization, while the algorithm class is now simply used as an attribute of the optimizer class.
