+++ midas/input_parser.py
@@ -23,6 +23,19 @@ class Input_Parser():
         
         self.parse_input_data()
         self.validate_input()
+    
+    def yaml_linereader(data,keyword,default):
+        """
+        #!TODO: write docstring.
+        
+        Written by Nicholas Rollins. 10/03/2024
+        """
+        try:
+            parsed_val = data[keyword]
+            #!TODO: validatation of value based on keyword
+        except:
+            parsed_val = default
+        return parsed_val
         
     def parse_input_data(self):
         """
@@ -33,32 +46,32 @@ class Input_Parser():
         ## Optimization Block ##
         info = self.file_settings['optimization']
         
-        self.methodology = info['methodology'].lower()
-        self.calculation_type = info['data_type'].lower() #!or external_model?
-        self.results_dir_name = info['results_directory_name']
-        self.population_size = info['population_size']
-        self.num_generations = info['number_of_generations']
-        self.parameters = info['objectives']
+        self.methodology = yaml_line_reader(info, 'methodology', None)
+        self.calculation_type = yaml_line_reader(info, 'data_type', None) #!or external_model?
+        self.results_dir_name = yaml_line_reader(info, 'results_directory_name', None)
+        self.population_size = yaml_line_reader(info, 'population_size', 1)
+        self.num_generations = yaml_line_reader(info, 'number_of_generations', 1)
+        self.parameters = yaml_line_reader(info, 'objectives', None)
         
-        self.genome = self.file_settings['decision_variables']['parameters'] #!TODO: add genome validation.
+        self.genome = yaml_line_reader(self.file_settings['decision_variables'], 'parameters', None) #!TODO: add genome validation.
         
         ## Algorithm Block ##
         info = self.file_settings['algorithm']
         
-        self.fitness = info['fitness'].lower()
-        self.selection = info['selection'].lower()
-        self.reproducer = info['reproducer'].lower()
-        self.mutation_type = info['mutation_type'].lower()
-        self.mutation_rate = float(info['mutation_rate'])
+        self.fitness = yaml_line_reader(info, 'fitness', None)
+        self.selection = yaml_line_reader(info, 'selection', None)
+        self.reproducer = yaml_line_reader(info, 'reproducer', None)
+        self.mutation_type = yaml_line_reader(info, 'mutation_type', None)
+        self.mutation_rate = yaml_line_reader(info, 'mutation_rate', None)
         
         ## Calculation Block ##
         info = self.file_settings['parcs_data']
         
-        self.xs_lib = info['xs_library_path']
-        self.power = float(info['power'])
-        self.flow = float(info['flow'])
-        self.inlet_temp = float(info['inlet_temperature'])
-        self.num_assemblies = float(info['number_assemblies'])
+        self.xs_lib = yaml_line_reader(info, 'xs_library_path', None)
+        self.power = yaml_line_reader(info, 'power', 100)
+        self.flow = yaml_line_reader(info, 'flow', None)
+        self.inlet_temp = yaml_line_reader(info, 'inlet_temperature', None)
+        self.num_assemblies = yaml_line_reader(info, 'number_assemblies', None)
         
         #!TODO: add core_dict and other values for PARCS/Simulate3
